Attempt to capture *test.go files properly with renovate
I think that the code generated by regex101.com was overzelous when
escaping backslashes in daccb94b6a0cb2890be9b2bb0bfe57760d8c94d4.

I think this fixes the issue, but it's hard to say defitively because I
see no way to run renovate on a local copy of a repository to see what
it is actually matching, unfortunately.

Signed-off-by: Shawn Wilsher <656602+sdwilsh@users.noreply.github.com>